+++ CHANGELOG.md
@@ -88,8 +88,11 @@
 
 * Use-after-free during shutdown when footclient and the client
   application exit simultaneously in a short window ([#2432][2432]).
+* Crash when a buggy compositor sends an `ime_enter()` event with a
+  `NULL` surface ([#2467][2467]).
 
 [2342]: https://codeberg.org/dnkl/foot/pulls/2432
+[2342]: https://codeberg.org/dnkl/foot/issues/2467
 
 
 ### Security